perf(thirdweb): gate thirdweb to desktop-only; mobile uses direct deposit / QR
External-wallet connect (thirdweb: MetaMask etc.) is a desktop-only feature.
Mounting the ThirdwebProvider on native + web-mobile forced its bundle/init and
Suspense onto the mobile first-paint path (and ran its hooks there). Gate it to
desktop so mobile no longer loads or runs thirdweb.

- app/_layout.tsx: mount LazyThirdwebProvider only when isDesktop (new `enabled`
  passthrough on the provider); on native + web-mobile the tree renders without
  it. Adds a desktop-only ThirdwebConnectionBridge that mirrors the connected
  wallet (address/status/account/wallet) into useDepositStore.
- Decouple always-mounted / mobile-rendered consumers from thirdweb hooks so
  they don't crash without the provider: useDepositOption and the vault EOA
  hooks (useDepositFromEOA/Eth/Fuse) now read the connection from the store
  instead of calling useActiveAccount/useActiveWallet. On mobile these stay
  disconnected (the connected-wallet deposit path is never used on mobile).
- Convert the existing native-vs-web splits into mobile-vs-desktop (isDesktop)
  splits so web-mobile never renders thirdweb-hook components:
  * CardDirectDepositModal.tsx → dispatcher (desktop=connect, mobile=QR/direct);
    extracted CardDirectDepositModal{Desktop,Mobile}.tsx.
  * DepositOptions.web.tsx → dispatcher; extracted DepositOptions{Desktop,Mobile}.tsx.
  * CardDepositOptions: "From External Wallet" now gated on isDesktop.

thirdweb desktop paths (external-wallet connect/deposit) are unchanged. All
reads still go through wagmi; getChain and the thirdweb client are provider-free.
